Fixes warnings from clang 1.7.
This fixes one case of format-string type mismatch with clang 1.7.  Since
the change simply coerces the type to the expected (and actual) value,
it should be reasonably safe.

TESTED:
Still builds without warnings on OSX 10.9 & 10.12, and Ubuntu 14.04.
Now builds without warnings on OSX 10.6.
No new warnings on OSX 10.5.

+++ libntp/socktoa.c
@@ -85,7 +85,7 @@ sockporttoa(
 		 (IS_IPV6(sock))
 		     ? "[%s]:%hu"
 		     : "%s:%hu",
-		 atext, SRCPORT(sock));
+		 atext, (unsigned short)SRCPORT(sock));
 	errno = saved_errno;
 
 	return buf;
